requirements.txt
pandas
akshare
pyyaml
fastapi
uvicorn[standard]
duckdb
pyarrow
python-dateutil
httpx
